Move MessageLocation mappers to message mappers file
