+++ common/src/main/java/com/ably/tracking/common/MessageMappers.kt
@@ -2,6 +2,7 @@ package com.ably.tracking.common
 
 import com.ably.tracking.Accuracy
 import com.ably.tracking.EnhancedLocationUpdate
+import com.ably.tracking.Location
 import com.ably.tracking.LocationUpdateType
 import com.ably.tracking.Resolution
 import com.google.gson.Gson
@@ -88,3 +89,29 @@ fun TripMetadata.toMessageJson(gson: Gson): String = gson.toJson(
         )
     )
 )
+
+fun Location.toMessage(): LocationMessage =
+    LocationMessage(
+        GeoJsonTypes.FEATURE,
+        LocationGeometry(GeoJsonTypes.POINT, listOf(longitude, latitude, altitude)),
+        LocationProperties(
+            accuracy,
+            bearing,
+            speed,
+            time.toDouble() / MILLISECONDS_PER_SECOND
+        )
+    )
+
+fun LocationMessage.toTracking(): Location =
+    Location(
+        longitude = geometry.coordinates[GEOMETRY_LONG_INDEX],
+        latitude = geometry.coordinates[GEOMETRY_LAT_INDEX],
+        altitude = geometry.coordinates[GEOMETRY_ALT_INDEX],
+        accuracy = properties.accuracyHorizontal,
+        bearing = properties.bearing,
+        speed = properties.speed,
+        time = (properties.time * MILLISECONDS_PER_SECOND).toLong()
+    )
+
+fun Message.getLocationMessages(gson: Gson): List<LocationMessage> =
+    gson.fromJson(data as String, Array<LocationMessage>::class.java).toList()
